+++ basicsr/archs/stylegan2_arch.py
@@ -7,8 +7,8 @@ from torch.nn import functional as F
 from basicsr.utils.registry import ARCH_REGISTRY
 
 try:
-    from basicsr.models.ops.fused_act import FusedLeakyReLU, fused_leaky_relu
-    from basicsr.models.ops.upfirdn2d import upfirdn2d
+    from basicsr.ops.fused_act import FusedLeakyReLU, fused_leaky_relu
+    from basicsr.ops.upfirdn2d import upfirdn2d
 except ImportError:
     print('Cannot import fused_act and upfirdn2d. Ignore this warning if '
           'they are not used. Otherwise install BasicSR with compiling them.')
